Fixes #5729 - content-host lookup by uuid
In the info, create, delete commands, non-id params can be used to
search for the resource. Currently the resource can not be found because
the command looks up the resource by its internal id instead of using the
uuid. This fix allows the commands to look up the resource by uuid.